fix: correct data types for energy consumption actual columns
Fixed type mismatch between database schema (BIGINT) and Rust structs (f64).
The consumption_l1_actual_w, consumption_l2_actual_w, and consumption_l3_actual_w
columns are BIGINT in the database, so the Rust types must be i64, not f64.

Changes:
- Updated EnergyLatest repository struct to use Option<i64> for actual_w fields
- Updated EnergyLatestResponse API model to match with Option<i64>
- Updated all tests to use i64 values instead of f64

This fixes the repository integration test failure where SQLx was rejecting
the type conversion from INT8 (BIGINT) to FLOAT8 (f64).

+++ applications/redpanda-sink/src/api/models/energy.rs
@@ -6,9 +6,9 @@ pub struct EnergyLatestResponse {
     pub ts: DateTime<Utc>,
     pub consumption_total_w: Option<f64>,
     pub consumption_total_actual_w: Option<i64>,
-    pub consumption_l1_actual_w: Option<f64>,
-    pub consumption_l2_actual_w: Option<f64>,
-    pub consumption_l3_actual_w: Option<f64>,
+    pub consumption_l1_actual_w: Option<i64>,
+    pub consumption_l2_actual_w: Option<i64>,
+    pub consumption_l3_actual_w: Option<i64>,
 }
 
 #[derive(Debug, Serialize)]
@@ -52,14 +52,14 @@ mod tests {
             ts: Utc::now(),
             consumption_total_w: Some(1500.0),
             consumption_total_actual_w: Some(1500),
-            consumption_l1_actual_w: Some(500.0),
-            consumption_l2_actual_w: Some(500.0),
-            consumption_l3_actual_w: Some(500.0),
+            consumption_l1_actual_w: Some(500),
+            consumption_l2_actual_w: Some(500),
+            consumption_l3_actual_w: Some(500),
         };
 
         assert!(response.consumption_total_w.is_some());
         assert_eq!(response.consumption_total_w.unwrap(), 1500.0);
-        assert_eq!(response.consumption_l1_actual_w.unwrap(), 500.0);
+        assert_eq!(response.consumption_l1_actual_w.unwrap(), 500);
     }
 
     #[test]
