+++ applications/redpanda-sink/src/repositories/energy.rs
@@ -32,9 +32,9 @@ pub struct EnergyLatest {
     pub ts: DateTime<Utc>,
     pub consumption_total_w: Option<f64>,
     pub consumption_total_actual_w: Option<i64>,
-    pub consumption_l1_actual_w: Option<f64>,
-    pub consumption_l2_actual_w: Option<f64>,
-    pub consumption_l3_actual_w: Option<f64>,
+    pub consumption_l1_actual_w: Option<i64>,
+    pub consumption_l2_actual_w: Option<i64>,
+    pub consumption_l3_actual_w: Option<i64>,
 }
 
 pub struct EnergyRepository;
